refactor(frontend): Use Modal component from react-bootstrap instead of the reimplementation
