+++ frontend/src/components/utils/Header.jsx
@@ -68,45 +68,9 @@ export default () => (
       </Col>
     </Row>
     <style jsx global>{`
-      @font-face {
-        font-family: "BentonSans";
-        src: url("/static/fonts/BentonSans-Medium.otf") format("opentype");
-      }
-
-      @font-face {
-        font-family: "BwSeidoRound";
-        font-weight: 500;
-        src: url("/static/fonts/Branding-with-Type-Bw-Seido-Round-Medium.otf")
-          format("opentype");
-      }
-
-      @font-face {
-        font-family: "BwSeidoRound";
+      .modal-body {
+        font-family: BentonSans;
         font-weight: 300;
-        src: url("/static/fonts/Branding-with-Type-Bw-Seido-Round-Light.otf")
-          format("opentype");
-      }
-
-      @font-face {
-        font-family: "BwSeidoRound";
-        font-weight: 400;
-        src: url("/static/fonts/Branding-with-Type-Bw-Seido-Round-Regular.otf")
-          format("opentype");
-      }
-
-      h1,
-      h2 {
-        font-family: BwSeidoRound;
-        font-weight: 500;
-        color: #24272a;
-      }
-
-      h1 {
-        font-size: 48px;
-      }
-
-      h2 {
-        font-size: 24px;
       }
 
       .header-container {
